Move `UserProfileView` tests into a separate file
